Fixed submission_date_uploaded field showing the wrong time
